agents/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
ddgs>=9.0.0
